# packages/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: StreamYolo API
  version: 0.0.1
  description: Room-first live creator platform — rooms, tips, private sessions, tokens, CCBill.
servers:
  - url: http://localhost:3001
    description: Local dev
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Session token — delivered via Set-Cookie on login; also accepted as Bearer header for native clients.

  # ─── Reusable parameters ───────────────────────────────────────────────────
  parameters:
    Cursor:
      in: query
      name: cursor
      schema: { type: string }
      description: Opaque pagination cursor from previous page.
    Limit:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Insufficient role or ownership
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string }
    PaginatedMeta:
      type: object
      required: [hasMore, nextCursor]
      properties:
        hasMore: { type: boolean }
        nextCursor:
          type: string
          nullable: true
...
    PublicUser:
      type: object
      required: [id, displayName, role, status, createdAt]
      properties:
        id: { type: string }
        displayName: { type: string }
        username: { type: string }
        role:
          type: string
          enum: [VIEWER, CREATOR, ADMIN, MODERATOR]
        status:
          type: string
          enum: [ACTIVE, SUSPENDED, DELETED]
        createdAt: { type: string, format: date-time }
    WalletSummary:
      type: object
      required: [tokenBalance, reservedTokenBalance, lifetimePurchasedTokens, lifetimeSpentTokens]
      properties:
        tokenBalance: { type: integer }
        reservedTokenBalance: { type: integer }
        lifetimePurchasedTokens: { type: integer }
        lifetimeSpentTokens: { type: integer }
...
    CreatorSummary:
      type: object
      required: [id, userId, displayName, isLive, user]
      properties:
        id: { type: string }
        userId: { type: string }
        displayName: { type: string }
        user: { $ref: '#/components/schemas/PublicUser' }
        avatarUrl:
          type: string
          nullable: true
        isLive: { type: boolean }
        status:
          type: string
          enum: [DRAFT, PENDING, ACTIVE, PAUSED, SUSPENDED, BANNED]
        bio:
          type: string
          nullable: true
        privateRateTokensPerMinute:
          type: integer
          nullable: true
        minPrivateMinutes:
          type: integer
          nullable: true
        privateViewerCamRequired:
          type: boolean
          nullable: true
        privateScreenShareAllowed:
          type: boolean
          nullable: true
        privateRulesText:
          type: string
          nullable: true
...
    RoomCategory:
      type: string
      enum: [MALE, FEMALE, COUPLES, TRANS]
    RoomTagDto:
      type: object
      required: [slug, label]
      properties:
        slug: { type: string }
        label: { type: string }
        group:
          type: string
          nullable: true
...
    RoomCard:
      type: object
      required: [id, title, status, visibility, creator, viewerCount, activePrivateSessionId, privateAvailable, tags, createdAt]
      properties:
        id: { type: string }
        title: { type: string }
        status:
          type: string
          enum: [DRAFT, LIVE, ENDED, SUSPENDED, HIDDEN]
        visibility:
          type: string
          enum: [PUBLIC, UNLISTED]
        category:
          type: string
          nullable: true
          description: MALE, FEMALE, COUPLES, or TRANS when set
        countryCode:
          type: string
          minLength: 2
          maxLength: 2
          nullable: true
        countryName:
          type: string
          nullable: true
        tags:
          type: array
          items: { $ref: '#/components/schemas/RoomTagDto' }
        creator: { $ref: '#/components/schemas/CreatorSummary' }
        thumbnailUrl:
          type: string
          nullable: true
        viewerCount: { type: integer }
        activePrivateSessionId:
          type: string
          nullable: true
        privateAvailable: { type: boolean }
        privateRateTokensPerMinute:
          type: integer
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        createdAt: { type: string, format: date-time }
    RoomDetail:
      type: object
      required: [id, title, status, visibility, creator, viewerCount, activePrivateSessionId, privateRateTokensPerMinute, minPrivateMinutes, privateViewerCamRequired, privateScreenShareAllowed, tags, createdAt]
      properties:
        id: { type: string }
        title: { type: string }
        status:
          type: string
          enum: [DRAFT, LIVE, ENDED, SUSPENDED, HIDDEN]
        visibility:
          type: string
          enum: [PUBLIC, UNLISTED]
        category:
          type: string
          nullable: true
          description: MALE, FEMALE, COUPLES, or TRANS when set
        countryCode:
          type: string
          minLength: 2
          maxLength: 2
          nullable: true
        countryName:
          type: string
          nullable: true
        tags:
          type: array
          items: { $ref: '#/components/schemas/RoomTagDto' }
        creator: { $ref: '#/components/schemas/CreatorSummary' }
        thumbnailUrl:
          type: string
          nullable: true
        coverUrl:
          type: string
          nullable: true
        viewerCount: { type: integer }
        activePrivateSessionId:
          type: string
          nullable: true
        privateRateTokensPerMinute: { type: integer }
        minPrivateMinutes: { type: integer }
        privateViewerCamRequired: { type: boolean }
        privateScreenShareAllowed: { type: boolean }
        privateRulesText:
          type: string
          nullable: true
        goal:
          type: object
          allOf:
            - { $ref: '#/components/schemas/RoomGoalDto' }
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        endedAt:
          type: string
          format: date-time
          nullable: true
        createdAt: { type: string, format: date-time }
    PrepareRoomInput:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        visibility:
          type: string
          enum: [PUBLIC, UNLISTED]
        thumbnailMediaId: { type: string }
        coverMediaId: { type: string }
        category: { $ref: '#/components/schemas/RoomCategory' }
        countryCode: { type: string, minLength: 2, maxLength: 2 }
        tagSlugs:
          type: array
          items: { type: string }
          maxItems: 10
        saveAsDefaults:
          type: boolean
          default: true
    GoLiveEligibility:
      type: object
      required: [canGoLive, missing]
      properties:
        canGoLive: { type: boolean }
        missing:
          type: array
          items:
            type: string
            enum:
              - CREATOR_NOT_APPROVED
              - ROOM_THUMBNAIL
              - ROOM_TITLE
              - ROOM_CATEGORY
              - ROOM_COUNTRY
              - PRIVATE_RULES_TEXT
              - PRIVATE_RATE
              - TIP_MENU
...
    RoomGoalDto:
      type: object
      required: [id, title, targetTokens, currentTokens]
      properties:
        id: { type: string }
        title: { type: string }
        targetTokens: { type: integer }
        currentTokens: { type: integer }

    # ── Creator menu items ────────────────────────────────────────────────────
...
    TipDto:
      type: object
      required: [id, roomId, fromUserId, toCreatorId, amountTokens, requestType, status, createdAt]
      properties:
        id: { type: string }
        roomId: { type: string }
        fromUserId: { type: string }
        toCreatorId: { type: string }
        amountTokens: { type: integer }
        requestType:
          type: string
          enum: [GENERAL, MENU_ITEM, CUSTOM, GOAL]
        menuItemId:
          type: string
          nullable: true
        requestText:
          type: string
          nullable: true
        status:
          type: string
          enum: [SENT, ACKNOWLEDGED, COMPLETED, DECLINED, REVERSED]
        createdAt: { type: string, format: date-time }
    CreateTipInput:
      type: object
      required: [amountTokens, requestType]
      properties:
        amountTokens: { type: integer, minimum: 1 }
        requestType:
          type: string
          enum: [GENERAL, MENU_ITEM, CUSTOM, GOAL]
        menuItemId: { type: string }
        requestText: { type: string, maxLength: 500 }

    # ── Token packs & Settings ───────────────────────────────────────────────
...
paths:
...
  /rooms:
    get:
      operationId: listRooms
      summary: List live public rooms
      security: []
      tags: [rooms]
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [LIVE, ENDED] }
        - in: query
          name: category
          schema:
            type: array
            items: { $ref: '#/components/schemas/RoomCategory' }
          style: form
          explode: true
        - in: query
          name: country
          schema:
            type: array
            items: { type: string, minLength: 2, maxLength: 2 }
          style: form
          explode: true
        - in: query
          name: tag
          schema:
            type: array
            items: { type: string }
          style: form
          explode: true
        - in: query
          name: q
          schema: { type: string }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated room list
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/RoomCard' }
                  meta: { $ref: '#/components/schemas/PaginatedMeta' }
...
  /creator/rooms:
    get:
      operationId: listCreatorRooms
      summary: List rooms created by the current creator
      tags: [creator-rooms]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated room list for the authenticated creator
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/RoomCard' }
                  meta: { $ref: '#/components/schemas/PaginatedMeta' }
        '401':
          $ref: '#/components/responses/Unauthorized'
  /creator/rooms/prepare:
    post:
      operationId: prepareRoom
      summary: Create or update the creator's draft room setup
      tags: [creator-rooms]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PrepareRoomInput' }
      responses:
        '200':
          description: Room prepared
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [room, goLiveEligibility]
                    properties:
                      room: { $ref: '#/components/schemas/RoomDetail' }
                      goLiveEligibility: { $ref: '#/components/schemas/GoLiveEligibility' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /creator/rooms/{roomId}:
    put:
      operationId: updateCreatorRoom
      summary: Update an existing room's details before or during broadcast
      tags: [creator-rooms]
      parameters:
        - in: path
          name: roomId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 3
                  maxLength: 100
                category: { $ref: '#/components/schemas/RoomCategory' }
                countryCode: { type: string, pattern: '^[A-Z]{2}$' }
                tagSlugs:
                  type: array
                  items: { type: string }
      responses:
        '200':
          description: Room updated
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [room]
                    properties:
                      room: { $ref: '#/components/schemas/RoomDetail' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /creator/rooms/{roomId}/end:
    post:
      operationId: endRoom
      summary: End the creator's live room
      tags: [creator-rooms]
      parameters:
        - in: path
          name: roomId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Room ended
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [room]
                    properties:
                      room: { $ref: '#/components/schemas/RoomDetail' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Creator profile ──────────────────────────────────────────────────────────
...
  /rooms/{roomId}:
    get:
      operationId: getRoom
      summary: Get a public room by ID
      security: []
      tags: [rooms]
      parameters:
        - in: path
          name: roomId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Room detail with VIP user list
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [room]
                    properties:
                      room: { $ref: '#/components/schemas/RoomDetail' }
                      vipUserIds:
                        type: array
                        items: { type: string }
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Chat ─────────────────────────────────────────────────────────────────────
...
  /rooms/{roomId}/tips:
    post:
      operationId: createTip
      summary: Send a tip in a room
      tags: [tips]
      parameters:
        - in: path
          name: roomId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTipInput' }
      responses:
        '201':
          description: Tip sent
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [tip, wallet]
                    properties:
                      tip: { $ref: '#/components/schemas/TipDto' }
                      wallet: { $ref: '#/components/schemas/WalletSummary' }
        '400':
          description: Insufficient tokens or invalid tip
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          $ref: '#/components/responses/Unauthorized'
  /creator/tips/{tipId}/acknowledge:
    post:
      operationId: acknowledgeTip
      summary: Creator acknowledges a tip
      tags: [tips]
      parameters:
        - in: path
          name: tipId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Tip acknowledged
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [tip]
                    properties:
                      tip: { $ref: '#/components/schemas/TipDto' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /creator/tips/{tipId}/complete:
    post:
      operationId: completeTip
      summary: Creator marks a tip request as completed
      tags: [tips]
      parameters:
        - in: path
          name: tipId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Tip completed
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [tip]
                    properties:
                      tip: { $ref: '#/components/schemas/TipDto' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Token packs ───────────────────────────────────────────────────────────────
...